Changes to testing files; isolated BTM_reValToVoltage() issue.

+++ components/bms/CANbus/inc/CANbus_TESTING_ONLY.h
@@ -323,7 +323,7 @@ typedef struct {
  * @param raw_reading The 16-bit reading from an LTC6811
  * @return Returns a properly scaled floating-point version of raw_reading
  */
-float BTM_regValToVoltage(uint16_t raw_reading)
+float BTM_regValToVoltage(float raw_reading)
 {
 	return raw_reading * BTM_VOLTAGE_CONVERSION_FACTOR;
 }
